fix: clean compile under -Dwarnings; bump rustls-webpki
Document the public MCP server types, make test_utils a module doc, and
resolve the clippy lints current stable surfaces (unnecessary casts,
is_multiple_of, collapsible ifs, field-reassign-with-default, range
loops) so the crate builds warning-free under RUSTFLAGS=-Dwarnings with
--all-features. Bump rustls-webpki 0.103.10 -> 0.103.13 to clear
RUSTSEC-2026-0098/0099/0104.

+++ src/capture/live.rs
@@ -143,8 +143,8 @@ pub static INVALID_PCAP_TIMESTAMPS: std::sync::atomic::AtomicU64 =
 /// (rate-limited), because silently substituted timestamps corrupt every
 /// downstream timing computation.
 fn pcap_ts_to_chrono(ts: libc::timeval) -> DateTime<Utc> {
-    let sec = ts.tv_sec as i64;
-    let usec = ts.tv_usec as i64;
+    let sec = ts.tv_sec;
+    let usec = ts.tv_usec;
     let converted = if (0..1_000_000).contains(&usec) {
         Utc.timestamp_opt(sec, usec as u32 * 1000).single()
     } else {
